requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-telegram-bot>=21.0
notion-client>=2.2.1
python-dotenv>=1.0.0
pydantic>=2.6.0
python-multipart>=0.0.6
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT
bcrypt>=4.0.0
requests
gunicorn 